Get rid of no ascii characters in log Use help functions for run_simulations.py Improve log messages
