+++ examples/monique/run_to_present/run_simulations.py
@@ -1,9 +1,9 @@
 import os
 import subprocess
 import sys
+from pathlib import Path
 import pandas as pd
 from idmtools.core import ItemType
-from filelock import FileLock
 import manifest
 import params
 import argparse
@@ -11,6 +11,9 @@ from idmtools.core.platform_factory import Platform
 from idmtools.entities.experiment import Experiment
 from idmtools.entities.templated_simulation import TemplatedSimulations
 from datetime import datetime
+
+sys.path.insert(0, str(Path(__file__).resolve().parent.parent))
+from run_simulation_helper.simulation_helper import update_suite_tracking, update_scenario_status_to_done
 
 tracking_file = os.path.join(manifest.CURRENT_DIR, "..", f"suite_tracking_{params.experiment_type}.csv")
 
@@ -39,40 +42,6 @@ def _pre_run(experiment: Experiment, **kwargs):
     from snt.utility.plugins import initialize_plugins
     initialize_plugins(**kwargs)
 
-def update_suite_tracking(experiment, suite, experiment_type, tracking_file, time_stamp):
-    lock_path = tracking_file + ".lock"
-    with FileLock(lock_path, timeout=30):
-        # Ensure the file exists with headers
-        if not os.path.exists(tracking_file):
-            with open(tracking_file, "w") as f:
-                f.write("suite_name,suite_id,experiment_name,experiment_id,experiment_type,start_time_stamp,end_time_stamp\n")
-
-        df = pd.read_csv(tracking_file)
-
-        match = df['suite_id'] == str(suite.id)
-
-        if match.any():
-            # Update existing row(s)
-            df.loc[match, 'experiment_name'] = experiment.name
-            df.loc[match, 'experiment_id'] = experiment.id
-            df.loc[match, 'end_time_stamp'] = time_stamp
-        else:
-            # Add new row if suite_id not found
-            new_row = {
-                'suite_name': suite.name,
-                'suite_id': suite.id,
-                'experiment_name': experiment.name,
-                'experiment_id': experiment.id,
-                'experiment_type': experiment_type,
-                'start_time_stamp': time_stamp,
-                'end_time_stamp': None
-            }
-            df = pd.concat([df, pd.DataFrame([new_row])], ignore_index=True)
-
-        # Save changes
-        df.to_csv(tracking_file, index=False)
-
-
 
 def _post_run(experiment, suite_id, **kwargs):
     """
@@ -90,13 +59,19 @@ def _post_run(experiment, suite_id, **kwargs):
         time_stamp = datetime.now().strftime("%Y%m%d_%H%M%S")
         update_suite_tracking(experiment, suite, params.experiment_type, tracking_file, time_stamp)
 
-        ssmt_script = os.path.join(manifest.CURRENT_DIR, "..", "analyzers", "post_ssmt.py")
-        ssmt_log = os.path.join(manifest.CURRENT_DIR, "..", "logs", f"ssmt_analyzer_{params.experiment_type}.log")
 
+        analyzer_script = os.path.join(manifest.CURRENT_DIR, "..", "analyzers", "post_ssmt.py")
+        # replace with local analyzer script if needed
+        #analyzer_script = os.path.join(manifest.CURRENT_DIR, "..", "analyzers", "post_analysis.py")
+
+        ssmt_log = os.path.join(manifest.CURRENT_DIR, "..", "logs", f"ssmt_analyzer_{params.experiment_type}_{params.scen_index}.log")
+
+        os.environ["NO_COLOR"] = "1"  # disable color in subprocess's log
+        # run ssmt (or local analyzer) subprocess in background
         with open(ssmt_log, "a+", encoding="utf-8") as log_file:
             subprocess.Popen([
                 sys.executable,
-                ssmt_script,
+                analyzer_script,
                 "--exp-id", str(experiment.id),
                 "--type", str(params.experiment_type),  # or hardcode 'future_projections' etc.
                 "--name", str(experiment.name)
@@ -105,7 +80,8 @@ def _post_run(experiment, suite_id, **kwargs):
                 stderr=subprocess.STDOUT,
                 encoding='utf-8'
             )
-        print(f"Launched SSMT analyzer for experiment: {experiment.name}--{experiment.id}")
+        print(f"{datetime.now().strftime('%Y-%m-%d %H:%M:%S,%f')[:-3]}: Launched SSMT analyzer for experiment: {experiment.name}--{experiment.id}")
+        print(f"{datetime.now().strftime('%Y-%m-%d %H:%M:%S,%f')[:-3]}: Analyzer_log: {ssmt_log}")
 
     print(f"EXPERIMENT_NAME: {experiment.name}")
     print(f"EXPERIMENT_ID: {experiment.id}")
@@ -139,23 +115,6 @@ def _config_experiment(**kwargs):
     experiment = Experiment.from_template(ts, name=experiment_name)
     return experiment
 
-def update_scenario_status_to_done(scenario_fname, scen_index):
-    """
-    Update scenario status to 'done' in scenario_fname. This is set to 'done' when the experiment is done.
-    Args:
-        scenario_fname:
-        scen_index:
-
-    Returns:
-
-    """
-    lockfile = scenario_fname + ".lock"
-    with FileLock(lockfile, timeout=30):
-        df = pd.read_csv(scenario_fname)
-        df.loc[scen_index, 'status'] = 'done'
-        df.to_csv(scenario_fname, index=False)
-        print(f"Updated scenario index {scen_index} to 'done' in {scenario_fname}")
-
 def run_experiment(**kwargs):
     """
     Get configured experiment and run.
@@ -172,12 +131,15 @@ def run_experiment(**kwargs):
     experiment = _config_experiment(**kwargs)
     _pre_run(experiment, **kwargs)
     if suite_id:
-        experiment.parent_id = suite_id
-    experiment.run(wait_until_done=True)
-    #experiment = platform.get_item("0aedf0fe-3567-f011-9f17-b88303912b51", ItemType.EXPERIMENT)  # debug purpose
-    update_scenario_status_to_done(params.scenario_fname, params.scen_index)
+        experiment.parent_id = suite_id  # Add experiment to suite
+    try:
+        experiment.run(wait_until_done=True)
+        # experiment = platform.get_item("cc510fbd-7b67-f011-9f17-b88303912b51", ItemType.EXPERIMENT)  # debug purpose
+        update_scenario_status_to_done(params.scenario_fname, params.scen_index)
 
-    _post_run(experiment, suite_id, **kwargs)
+        _post_run(experiment, suite_id, **kwargs)
+    except Exception as e:
+        print(f"Experiment run failed: {e}")
 
 
 if __name__ == "__main__":
@@ -207,10 +169,10 @@ if __name__ == "__main__":
     # Determine warning level
     show_warnings_once = str_to_bool_none(args.show_warnings_once)
 
-    time_stamp = datetime.now().strftime("%Y%m%d_%H%M%S")
+    time_stamp = datetime.now().strftime("%Y-%m-%d %H:%M:%S,%f")[:-3]
     print("\n\n")
-    print("============================New Experiment==============================")
-    print(f'{time_stamp}: Start experiment with {args.suite_id}')
+    print("======================================New Experiment========================================")
+    print(f'{time_stamp}: Start experiment with suite_id: {args.suite_id}')
     platform = Platform('CALCULON')
     # If you don't have Eradication, un-comment out the following to download Eradication
     # import emod_malaria.bootstrap as dtk
